Move task heap into new file, make implementation cleaner.
